Added .gitignore entry to ignore emacs files.
On index.html added references to the new files I created to learn to use jaws.js

Created some "FAP-*" files to play around with.  They are a copy from existing states and objects, and adapated to learn how to use them.

I also added an entry to the MenuState to allo for the newly created FAPState.